Move 'Callbacks' file to 'handlers' package
