+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/dispatcher/HandlersDsl.kt
@@ -1,26 +1,5 @@
 package com.github.kotlintelegrambot.dispatcher
 
-import com.github.kotlintelegrambot.HandleAnimationUpdate
-import com.github.kotlintelegrambot.HandleAudioUpdate
-import com.github.kotlintelegrambot.HandleCallbackQuery
-import com.github.kotlintelegrambot.HandleCommand
-import com.github.kotlintelegrambot.HandleContact
-import com.github.kotlintelegrambot.HandleDice
-import com.github.kotlintelegrambot.HandleDocumentUpdate
-import com.github.kotlintelegrambot.HandleError
-import com.github.kotlintelegrambot.HandleGameUpdate
-import com.github.kotlintelegrambot.HandleInlineQuery
-import com.github.kotlintelegrambot.HandleLocation
-import com.github.kotlintelegrambot.HandleMessage
-import com.github.kotlintelegrambot.HandleNewChatMembers
-import com.github.kotlintelegrambot.HandlePhotosUpdate
-import com.github.kotlintelegrambot.HandlePollAnswer
-import com.github.kotlintelegrambot.HandleStickerUpdate
-import com.github.kotlintelegrambot.HandleText
-import com.github.kotlintelegrambot.HandleUpdate
-import com.github.kotlintelegrambot.HandleVideoNoteUpdate
-import com.github.kotlintelegrambot.HandleVideoUpdate
-import com.github.kotlintelegrambot.HandleVoiceUpdate
 import com.github.kotlintelegrambot.dispatcher.handlers.CallbackQueryHandler
 import com.github.kotlintelegrambot.dispatcher.handlers.ChannelHandler
 import com.github.kotlintelegrambot.dispatcher.handlers.CheckoutHandler
@@ -28,6 +7,27 @@ import com.github.kotlintelegrambot.dispatcher.handlers.CommandHandler
 import com.github.kotlintelegrambot.dispatcher.handlers.ContactHandler
 import com.github.kotlintelegrambot.dispatcher.handlers.DiceHandler
 import com.github.kotlintelegrambot.dispatcher.handlers.ErrorHandler
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleAnimationUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleAudioUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleCallbackQuery
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleCommand
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleContact
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleDice
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleDocumentUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleError
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleGameUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleInlineQuery
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleLocation
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleMessage
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleNewChatMembers
+import com.github.kotlintelegrambot.dispatcher.handlers.HandlePhotosUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandlePollAnswer
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleStickerUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleText
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleVideoNoteUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleVideoUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleVoiceUpdate
 import com.github.kotlintelegrambot.dispatcher.handlers.InlineQueryHandler
 import com.github.kotlintelegrambot.dispatcher.handlers.LocationHandler
 import com.github.kotlintelegrambot.dispatcher.handlers.MessageHandler
