+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/dispatcher/handlers/ChannelHandler.kt
@@ -1,6 +1,5 @@
 package com.github.kotlintelegrambot.dispatcher.handlers
 
-import com.github.kotlintelegrambot.HandleUpdate
 import com.github.kotlintelegrambot.entities.Update
 
 class ChannelHandler(handleUpdate: HandleUpdate) : Handler(handleUpdate) {
