+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/dispatcher/handlers/DiceHandler.kt
@@ -1,8 +1,6 @@
 package com.github.kotlintelegrambot.dispatcher.handlers
 
 import com.github.kotlintelegrambot.Bot
-import com.github.kotlintelegrambot.HandleDice
-import com.github.kotlintelegrambot.HandleUpdate
 import com.github.kotlintelegrambot.entities.Update
 
 internal class DiceHandler(handleDice: HandleDice) : Handler(HandleDiceProxy(handleDice)) {
@@ -12,7 +10,8 @@ internal class DiceHandler(handleDice: HandleDice) : Handler(HandleDiceProxy(han
     override fun checkUpdate(update: Update): Boolean = update.message?.dice != null
 }
 
-private class HandleDiceProxy(private val handleDice: HandleDice) : HandleUpdate {
+private class HandleDiceProxy(private val handleDice: HandleDice) :
+    HandleUpdate {
     override fun invoke(bot: Bot, update: Update) {
         val message = update.message
         val dice = message?.dice
