+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/dispatcher/handlers/Handler.kt
@@ -1,6 +1,5 @@
 package com.github.kotlintelegrambot.dispatcher.handlers
 
-import com.github.kotlintelegrambot.HandleUpdate
 import com.github.kotlintelegrambot.entities.Update
 
 abstract class Handler(val handlerCallback: HandleUpdate) {
