+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/dispatcher/handlers/media/AnimationHandler.kt
@@ -1,6 +1,6 @@
 package com.github.kotlintelegrambot.dispatcher.handlers.media
 
-import com.github.kotlintelegrambot.HandleAnimationUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleAnimationUpdate
 import com.github.kotlintelegrambot.entities.Update
 import com.github.kotlintelegrambot.entities.files.Animation
