+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/dispatcher/handlers/media/AudioHandler.kt
@@ -1,6 +1,6 @@
 package com.github.kotlintelegrambot.dispatcher.handlers.media
 
-import com.github.kotlintelegrambot.HandleAudioUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleAudioUpdate
 import com.github.kotlintelegrambot.entities.Update
 import com.github.kotlintelegrambot.entities.files.Audio
