+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/dispatcher/handlers/media/VideoHandler.kt
@@ -1,6 +1,6 @@
 package com.github.kotlintelegrambot.dispatcher.handlers.media
 
-import com.github.kotlintelegrambot.HandleVideoUpdate
+import com.github.kotlintelegrambot.dispatcher.handlers.HandleVideoUpdate
 import com.github.kotlintelegrambot.entities.Update
 import com.github.kotlintelegrambot.entities.files.Video
