+++ telegram/src/test/kotlin/com/github/kotlintelegrambot/dispatcher/handlers/ContactHandlerTest.kt
@@ -4,7 +4,6 @@ import anyContact
 import anyMessage
 import anyUpdate
 import com.github.kotlintelegrambot.Bot
-import com.github.kotlintelegrambot.HandleContact
 import io.mockk.mockk
 import io.mockk.verify
 import junit.framework.TestCase.assertFalse
